+++ src/submissions/tools/__init__.py
@@ -261,3 +261,4 @@ def massage_common_reagents(reagent_name:str):
     reagent_name = reagent_name.replace("µ", "u")
     return reagent_name
         
+